move power to separate file
